# apps/whatsnxt-bff/swagger.json
{
  "swagger": "2.0",
  "paths": {
...
    "/api/v1/comment/{id}/toggleLike": {
      "post": {
        "summary": "/api/v1/comment/{id}/toggleLike",
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {},
        "tags": []
      }
    },
    "/api/v1/comment/{id}/toggleDislike": {
      "post": {
        "summary": "/api/v1/comment/{id}/toggleDislike",
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {},
        "tags": []
      }
    },
...
  },
...
  "info": {
    "title": "whatsnxt-courses-be",
    "version": "1.0.0",
    "description": "Specification JSONs: [v2](/api-spec/v2), [v3](/api-spec/v3).\n\nwhatsnxt backend"
  }
}